pyproject.toml
[tool.poetry]
name = "vacation_manager"
version = "0.1.0"
description = ""
authors = ["ktomi96 <kokai.tamas.96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.6.2"
python-dotenv = "^0.21.0"
Jinja2 = "^3.1.2"
MarkupSafe = "^2.1.1"
oauthlib = "^3.2.0"
requests = "^2.28.1"
pyOpenSSL = "^22.0.0"
Flask-WTF = "^1.0.1"
Flask-Mail = "^0.9.1"
livereload = "^2.6.3"
email-validator = "^1.2.1"



[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
